research-assistant-rl/requirements.txt
requests
numpy
matplotlib
arxiv
pandas
scipy
pinecone-client
sentence-transformers
streamlit
anthropic
networkx
plotly